Fix bug in euler.primes function prime_sieve
The function did not correctly run up to the bound given, it instead ran
up to the number - 1. Some problems assumed this behavior, and others
did not. The function is now expected to generate all primes up to and
including the number given as the bound on the sieve.

Previous solutions which assumed incorrect behavior have been updated.