Fix fibonacci_binet()
Previously, it produced incorrect results for all n due to lack of
rounding. It now produces the right results up to n = 70.

For n > 70, floating point error breaks the algorithm. For n > 70 an
ArithmeticError is raised, although this behavior can be turned off with
an optional argument to the function. This function is still useful for
n > 70 when only the leading digits of the Fibonacci number are needed
(for example, in Problem 104).